chore: refactor test case
